+++ packages/next/pages/user/[...params].js
@@ -2,6 +2,7 @@ import { withCommonProps } from "next-common/lib";
 import { backendApi } from "next-common/services/nextApi";
 import Profile from "next-common/components/profile";
 import { fetchOpenGovTracksProps } from "next-common/services/serverSide";
+import { fetchProfileTreasuryProps } from "next-common/services/serverSide/prefile/treasury";
 import { tryConvertToSubstrateAddress } from "next-common/utils/mixedChainUtil";
 
 export default Profile;
@@ -18,6 +19,7 @@ export const getServerSideProps = withCommonProps(async (context) => {
     backendApi.fetch(`users/${maybeAddress}`),
   ]);
   const tracksProps = await fetchOpenGovTracksProps();
+  const treasuryProps = await fetchProfileTreasuryProps(maybeAddress);
 
   return {
     props: {
@@ -26,6 +28,7 @@ export const getServerSideProps = withCommonProps(async (context) => {
       user: user ?? {},
       route: context.query?.params?.slice(1)?.join("/") ?? "",
       ...tracksProps,
+      ...treasuryProps,
     },
   };
 });